Split out suggest, more cleanup
